docs: fix review suggestion
