+++ docs/I18N.md
@@ -4,7 +4,8 @@
 
 - 中文文档：`docs/src/`（不含 `en/`）
 - 英文文档：`docs/src/en/`（与中文路径一一对应）
-- Demo：`docs/demos/`（中英文版本分开）
+- 中文 Demo：`docs/demos/`（默认）
+- 英文 Demo：`docs/demos/en/`（与中文路径一一对应）
 - 图片：`docs/src/public/`（中英文共用）
 - 中文导航：`docs/.vitepress/config/zh-theme.ts`
 - 英文导航：`docs/.vitepress/config/en-theme.ts`
@@ -15,8 +16,8 @@
 
 1. 编写中文 `.md`
 2. 有交互示例时，在 `docs/demos/` 新增 `.vue`
-   - 中文版本：`my-feature.vue`
-   - 英文版本（如有中文内容）：`my-feature_en.vue`
+   - 中文版本：`docs/demos/chat/my-feature.vue`
+   - 英文版本（如有中文内容）：`docs/demos/en/chat/my-feature.vue`
 3. 有图片时，放入 `docs/src/public/`
 4. 在 `zh-theme.ts` 的 sidebar 中增加链接：
 
@@ -37,8 +38,8 @@
 
 ### 命名规则
 
-- 中文版本：`xxx.vue`（默认）
-- 英文版本：`xxx_en.vue`
+- 中文版本：`docs/demos/xxx.vue`（默认）
+- 英文版本：`docs/demos/en/xxx.vue`（路径与中文一一对应，文件名相同）
 
 ### 使用方式
 
@@ -49,7 +50,7 @@
 <demo vue="../../../demos/chat/history.vue" />
 
 <!-- 英文文档 -->
-<demo vue="../../../../demos/chat/history_en.vue" />
+<demo vue="../../../../demos/en/chat/history.vue" />
 ```
 
 ### 需要国际化的内容
@@ -73,14 +74,14 @@ Demo 文件中以下内容需要翻译：
 
 ## 英文文档注意事项
 
-从中文复制到英文时，`<demo>` 和图片的相对路径需**多加一层 `../`**：
+从中文复制到英文时，`<demo>` 和图片的相对路径需**多加一层 `../`**，并将 demo 路径指向 `demos/en/`：
 
 ```markdown
 <!-- 中文 -->
 <demo vue="../../../demos/chat/my-feature.vue" />
 
 <!-- 英文 -->
-<demo vue="../../../../demos/chat/my-feature_en.vue" />
+<demo vue="../../../../demos/en/chat/my-feature.vue" />
 ```
 
 文档内互相引用用相对路径即可，写法与中文相同。  
@@ -93,4 +94,4 @@ sidebar / nav 配置中，英文链接需带 `/en` 前缀。
 | 改正文 | 更新 `src/en/` 下对应文件 |
 | 重命名 / 移动 | 同步移动英文文件，更新两个 `theme` 配置 |
 | 删除 | 删除英文镜像文件，更新两个 `theme` 配置 |
-| 新增/修改 Demo | 同步更新 `_en.vue` 版本（如有中文内容） |
+| 新增/修改 Demo | 同步更新 `demos/en/` 下对应文件（如有中文内容） |
