+++ docs/demos/en/angular/renderer/custom-actions-open-page.vue
@@ -4,7 +4,7 @@
 
 <script setup lang="ts">
 import { ref } from 'vue';
-import GenuiRenderer from './adapter.vue';
+import GenuiRenderer from '../../../angular/renderer/adapter.vue';
 
 const generating = ref(false);
 const content = ref({
