+++ docs/demos/en/chat/components/assistant-footer.vue
