+++ docs/demos/en/chat/custom-components.vue
@@ -4,7 +4,7 @@
 
 <script setup lang="ts">
 import { GenuiChat } from '@opentiny/genui-sdk-vue';
-import UserProfile from './components/user-profile.vue';
+import UserProfile from '../../chat/components/user-profile.vue';
 
 const url = 'https://your-chat-backend/api';
