+++ docs/demos/en/chat/custom-fetch.vue
@@ -53,4 +53,3 @@ function generateRequestId(): string {
   return `req-${Date.now()}-${Math.random().toString(36).substr(2, 9)}`;
 }
 </script>
-
