+++ docs/demos/en/chat/custom-snippets.vue
@@ -223,4 +223,3 @@ const messages = [
   },
 ];
 </script>
-
