+++ docs/demos/en/chat/footer-toolbar.vue
@@ -1,27 +1,53 @@
 <template>
-  <div class="app">
-    <header class="app-header">
-      <h1>GenUI Chat</h1>
-      <div class="header-actions">
-        <button @click="toggleTheme" class="theme-toggle">
-          {{ themeIcon }}
-        </button>
-      </div>
-    </header>
-
-    <GenuiConfigProvider :theme="currentTheme" id="main-chat">
-      <GenuiChat :url="url" model="deepseek-v3.2" :messages="messages" />
-    </GenuiConfigProvider>
-  </div>
+  <GenuiChat :url="url" :roles="roles" :messages="messages" />
 </template>
 
 <script setup lang="ts">
-import { ref, computed } from 'vue';
-import { GenuiConfigProvider, GenuiChat } from '@opentiny/genui-sdk-vue';
+import { GenuiChat } from '@opentiny/genui-sdk-vue';
+import AssistantFooter from './components/assistant-footer.vue';
+import UserFooter from './components/user-footer.vue';
 
 const url = 'https://your-chat-backend/api';
 
+const roles = {
+  assistant: {
+    slots: {
+      trailer: AssistantFooter,
+    },
+  },
+  user: {
+    slots: {
+      trailer: UserFooter,
+    },
+  },
+};
+
 const messages = [
+  {
+    role: 'user',
+    content: 'Generate a button',
+  },
+  {
+    role: 'assistant',
+    content: '',
+    messages: [
+      {
+        type: 'schema-card',
+        content: JSON.stringify({
+          componentName: 'Page',
+          children: [
+            {
+              componentName: 'TinyButton',
+              props: {
+                type: 'primary',
+                text: 'Primary Button',
+              },
+            },
+          ],
+        }),
+      },
+    ],
+  },
   {
     role: 'user',
     content: 'Search for train tickets',
@@ -92,52 +118,4 @@ const messages = [
     ],
   },
 ];
-
-const currentTheme = ref<'dark' | 'lite' | 'light'>('light');
-
-const themeIcon = computed(() => {
-  const icons = {
-    light: '☀️',
-    dark: '🌙',
-  };
-  return icons[currentTheme.value];
-});
-
-function toggleTheme() {
-  currentTheme.value = currentTheme.value === 'dark' ? 'light' : 'dark';
-}
 </script>
-
-<style scoped>
-.app {
-  display: flex;
-  flex-direction: column;
-  height: 800px;
-}
-
-.app-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 16px;
-  border-bottom: 1px solid #eee;
-}
-
-.app-header h1 {
-  margin: 0;
-  font-size: 20px;
-}
-
-.theme-toggle {
-  padding: 8px 12px;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  background: white;
-  cursor: pointer;
-  font-size: 20px;
-}
-
-.theme-toggle:hover {
-  background: #f5f5f5;
-}
-</style>
