+++ docs/demos/en/chat/image-upload.vue
