+++ docs/demos/en/config-provider/custom-theme.vue
