+++ docs/demos/en/config-provider/schema-renderer-theme.vue
