+++ docs/demos/en/renderer/custom-actions-open-page.vue
