+++ docs/demos/en/renderer/state.vue
@@ -1,5 +1,5 @@
 <template>
-  <GenuiRenderer :content="content" :generating="generating" :customActions="customActions" />
+  <GenuiRenderer :content="content" :generating="generating" :state="historyState" :customActions="customActions" />
 </template>
 
 <script setup lang="ts">
@@ -7,14 +7,16 @@ import { ref } from 'vue';
 import { GenuiRenderer } from '@opentiny/genui-sdk-vue';
 
 const generating = ref(false);
+
+// Schema restored from history
 const content = ref({
+  componentName: 'Page',
   state: {
     formData: {
       name: '',
-      age: '',
+      age: null,
     },
   },
-  componentName: 'Page',
   children: [
     {
       componentName: 'TinyForm',
@@ -23,7 +25,7 @@ const content = ref({
           type: 'JSExpression',
           value: 'this.state.formData',
         },
-        labelPosition: 'top',
+        'label-position': 'top',
       },
       children: [
         {
@@ -36,7 +38,6 @@ const content = ref({
             {
               componentName: 'TinyInput',
               props: {
-                placeholder: 'Enter name',
                 modelValue: {
                   type: 'JSExpression',
                   model: true,
@@ -56,7 +57,6 @@ const content = ref({
             {
               componentName: 'TinyInput',
               props: {
-                placeholder: 'Enter age',
                 modelValue: {
                   type: 'JSExpression',
                   model: true,
@@ -66,49 +66,34 @@ const content = ref({
             },
           ],
         },
-        {
-          componentName: 'TinyFormItem',
-          props: {
-            label: '',
-          },
-          children: [
-            {
-              componentName: 'TinyButton',
-              props: {
-                type: 'primary',
-                text: 'Show Form Content',
-                onClick: {
-                  type: 'JSFunction',
-                  value: "function() { this.callAction('showNotification', { title: 'Form Content' }); }",
-                },
-              },
-            },
-          ],
-        },
       ],
+    },
+    {
+      componentName: 'TinyButton',
+      props: {
+        text: 'View State',
+        onClick: {
+          type: 'JSFunction',
+          value: "function() { this.callAction('getState'); }",
+        },
+      },
     },
   ],
 });
 
+// State restored from history (merged into global state on initialization)
+const historyState = {
+  formData: {
+    name: 'John Doe',
+    age: 30,
+  },
+};
+
 const customActions = {
-  showNotification: {
-    name: 'showNotification',
-    description: 'Display a notification with the form content',
+  getState: {
     execute: (params: any, context: Record<string, any>) => {
       const state = context.state;
-      const message = JSON.stringify(state);
-
-      alert(`${params.title}: ${message}`);
-    },
-    parameters: {
-      type: 'object',
-      properties: {
-        title: {
-          type: 'string',
-          description: 'Notification title',
-        },
-      },
-      required: ['title'],
+      alert(`History State:\n${JSON.stringify(state, null, 2)}`);
     },
   },
 };
