+++ docs/demos/renderer/components/user-profile.vue
@@ -72,4 +72,3 @@ defineProps<{
   color: #666;
 }
 </style>
-
