+++ docs/src/en/examples/config-provider/theme.md
@@ -4,8 +4,8 @@
 
 ## Theme for GenuiChat
 
-<demo vue="../../../../demos/config-provider/theme_en.vue" />
+<demo vue="../../../../demos/en/config-provider/theme.vue" />
 
 ## Theme for GenuiRenderer
 
-<demo vue="../../../../demos/config-provider/schema-renderer-theme_en.vue" />
+<demo vue="../../../../demos/en/config-provider/schema-renderer-theme.vue" />
